# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dasma Pro - Digital Wedding & Event Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/{eventId}/guests:
    get:
      operationId: listGuests
      tags: [guests]
      summary: List guests for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: number
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of guests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guest"

    post:
      operationId: createGuest
      tags: [guests]
      summary: Add a guest to an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestInput"
      responses:
        "201":
          description: Created guest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
  /events/{eventId}/guests/import:
    post:
      operationId: importGuests
      tags: [guests]
      summary: Bulk import guests from CSV data
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestImport"
      responses:
        "201":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /events/{eventId}/guests/{guestId}:
    get:
      operationId: getGuest
      tags: [guests]
      summary: Get a single guest
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: number
        - name: guestId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Guest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"

    patch:
      operationId: updateGuest
      tags: [guests]
      summary: Update a guest
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: number
        - name: guestId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestUpdate"
      responses:
        "200":
          description: Updated guest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"

    delete:
      operationId: deleteGuest
      tags: [guests]
      summary: Delete a guest
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: number
        - name: guestId
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─── TABLES ───────────────────────────────────────────────────────────────
...
  /events/{eventId}/hall:
    get:
      operationId: getHallLayout
      tags: [hall]
      summary: Get hall layout for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Hall layout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HallLayout"

    put:
      operationId: saveHallLayout
      tags: [hall]
      summary: Save hall layout
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HallLayoutInput"
      responses:
        "200":
          description: Saved layout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HallLayout"

  # ─── INVITATIONS ──────────────────────────────────────────────────────────
...
  /events/{eventId}/guests/lookup:
    get:
      operationId: lookupGuest
      tags: [checkins]
      summary: Find guest by name (table finder)
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: number
        - name: name
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching guests with table info
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GuestLocation"

  # ─── ADMIN ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Guest:
      type: object
      properties:
        id:
          type: number
        eventId:
          type: number
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        partySize:
          type: number
        category:
          type: string
          enum: [family, friends, colleagues, other]
        status:
          type: string
          enum: [pending, invited, confirmed, declined, checked_in]
        tableId:
          type: ["number", "null"]
        seatNumber:
          type: ["number", "null"]
        rsvpToken:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - eventId
        - firstName
        - lastName
        - partySize
        - category
        - status
        - rsvpToken
        - createdAt
    GuestInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        email:
          type: string
        partySize:
          type: number
        category:
          type: string
          enum: [family, friends, colleagues, other]
        notes:
          type: string
      required:
        - firstName
        - lastName
        - partySize
        - category
    GuestUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        email:
          type: string
        partySize:
          type: number
        category:
          type: string
          enum: [family, friends, colleagues, other]
        status:
          type: string
          enum: [pending, invited, confirmed, declined, checked_in]
        notes:
          type: string
    GuestImport:
      type: object
      properties:
        guests:
          type: array
          items:
            $ref: "#/components/schemas/GuestInput"
      required:
        - guests
    ImportResult:
      type: object
      properties:
        imported:
          type: number
        failed:
          type: number
        errors:
          type: array
          items:
            type: string
      required:
        - imported
        - failed
        - errors
...
    HallLayout:
      type: object
      properties:
        id:
          type: number
        eventId:
          type: number
        width:
          type: number
        height:
          type: number
        elements:
          type: array
          items:
            $ref: "#/components/schemas/HallElement"
        updatedAt:
          type: string
      required:
        - id
        - eventId
        - width
        - height
        - elements
        - updatedAt
    HallElement:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [table, stage, dance_floor, dj, band, buffet, emergency_exit, entrance]
        x:
          type: number
        y:
          type: number
        width:
          type: number
        height:
          type: number
        rotation:
          type: number
        label:
          type: ["string", "null"]
        tableId:
          type: ["number", "null"]
        shape:
          type: ["string", "null"]
      required:
        - id
        - type
        - x
        - y
        - width
        - height
        - rotation
    HallLayoutInput:
      type: object
      properties:
        width:
          type: number
        height:
          type: number
        elements:
          type: array
          items:
            $ref: "#/components/schemas/HallElement"
      required:
        - width
        - height
        - elements
...
    GuestLocation:
      type: object
      properties:
        guestId:
          type: number
        firstName:
          type: string
        lastName:
          type: string
        tableName:
          type: ["string", "null"]
        tableId:
          type: ["number", "null"]
        seatNumber:
          type: ["number", "null"]
        status:
          type: string
      required:
        - guestId
        - firstName
        - lastName
        - status
...