# lib/api-zod/src/generated/types/guestUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Dasma Pro - Digital Wedding & Event Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { GuestUpdateCategory } from './guestUpdateCategory';
import type { GuestUpdateStatus } from './guestUpdateStatus';

export interface GuestUpdate {
  firstName?: string;
  lastName?: string;
  phone?: string;
  email?: string;
  partySize?: number;
  category?: GuestUpdateCategory;
  status?: GuestUpdateStatus;
  notes?: string;
}